Fix GError handling to prevent segfaults.
The error reporting caused a segfault on OSX
in g_object_unref. Replaced it by g_error_free, which is used
in the respective examples in the documentation:
https://developer.gnome.org/glib/stable/glib-Error-Reporting.html

Explicitly initialized GError to NULL.

At the end, I still don't get the proper error message in case of
"file not found" (err->message is an empty string) but at least
it does not crash afterwards.