Makefile.am: add bootstrap script and README.md to EXTRA_DIST.
The release tarballs currently don't contain README.md, because the
"dist" makefile target doesn't know about it. This is easily fixed by
adding "README.md" to the existing EXTRA_DIST variable in Makefile.am,
which is nothing other than a list of "extra" files to be included in
the release tarball.

The "bootstrap" script for autotools is also convenient to have in
the releases, so it has been added as well.